AlertBanner (and LateDepartmentsBanner) style tweaks
